CB-1815. Factor out grpc-common module.
With these changes, there is a new grpc-common module
for classes and proto files that will be shared across
our various modules that use GRPC.
