+++ auth-connector/build.gradle
@@ -15,6 +15,7 @@ buildscript {
 
 dependencies {
   compile project(':common')
+  compile project(':grpc-common')
 
   compile group: 'org.springframework.boot',            name: 'spring-boot-starter-jersey',     version: '2.0.3.RELEASE'
   compile group: 'org.springframework.security.oauth',  name: 'spring-security-oauth2',         version: springOauthVersion
