+++ grpc-common/build.gradle
@@ -1,0 +1,48 @@
+apply plugin: 'com.google.protobuf'
+
+buildscript {
+  repositories {
+    mavenLocal()
+    mavenCentral()
+    maven { url = "$repoUrl" }
+    jcenter()
+  }
+  dependencies {
+    classpath 'com.google.protobuf:protobuf-gradle-plugin:0.8.8'
+  }
+}
+
+dependencies {
+  implementation group: 'org.slf4j',                             name: 'slf4j-api',               version: slf4jApiVersion
+
+  implementation ("io.grpc:grpc-all:1.19.0") {
+    exclude group: 'org.mockito'
+  }
+
+  testImplementation group: 'org.mockito',                     name: 'mockito-core', version: mockitoVersion
+  testImplementation group: 'junit', name: 'junit', version: '4.12'
+}
+
+checkstyle {
+  checkstyleMain.exclude '**/generated/**'
+}
+
+protobuf {
+  protoc {
+    // The version of protoc must match protobuf-java. If you don't depend on
+    // protobuf-java directly, you will be transitively depending on the
+    // protobuf-java version that grpc depends on.
+    artifact = "com.google.protobuf:protoc:3.6.1"
+  }
+  plugins {
+    grpc {
+      artifact = 'io.grpc:protoc-gen-grpc-java:1.12.0'
+    }
+  }
+  generateProtoTasks {
+    all()*.plugins {
+      grpc {}
+    }
+  }
+  generatedFilesBaseDir = "$projectDir/src/generated"
+}
