+++ settings.gradle
@@ -33,6 +33,7 @@ include 'template-manager-cmtemplate'
 include 'cluster-api'
 include 'cluster-ambari'
 include 'cluster-cm'
+include 'grpc-common'
 include 'auth-connector'
 include 'flow'
 include 'freeipa'
